Spike Working Map + Apollo Request
This commit spikes a working/loading Map from Mapbox along with the initial request from the GraphQL API (in development) so that we can see the data make its way to the client.

We refactor the Map class component into a more modern React hook because Apollo now relies on React Hooks and the class based components aren’t compatible with it (at least not in the way they documented).